+++ src/ResourceManager/Network/Commands.Network/Models/PSSubnet.cs
@@ -44,7 +44,11 @@ namespace Microsoft.Azure.Commands.Network.Models
         [JsonProperty(Order = 1)]
         public List<PSServiceEndpoint> ServiceEndpoints { get; set; }
 
+        [JsonProperty(Order = 1)]
         public List<PSServiceEndpointPolicy> ServiceEndpointPolicies { get; set; }
+
+        [JsonProperty(Order = 1)]
+        public List<PSInterfaceEndpoint> InterfaceEndpoints { get; set; }
 
         [JsonProperty(Order = 1)]
         [Ps1Xml(Target = ViewControl.Table)]
@@ -85,7 +89,27 @@ namespace Microsoft.Azure.Commands.Network.Models
             return !string.IsNullOrEmpty(this.Name);
         }
 
-         [JsonIgnore]
+        public bool ShouldSerializeServiceEndpointPolicies()
+        {
+            return !string.IsNullOrEmpty(this.Name);
+        }
+
+        public bool ShouldSerializeServiceEndpoints()
+        {
+            return !string.IsNullOrEmpty(this.Name);
+        }
+
+        public bool ShouldSerializeResourceNavigationLinks()
+        {
+            return !string.IsNullOrEmpty(this.Name);
+        }
+
+        public bool ShouldSerializeInterfaceEndpoints()
+        {
+            return !string.IsNullOrEmpty(this.Name);
+        }
+
+        [JsonIgnore]
         public string ServiceEndpointText
         {
             get { return JsonConvert.SerializeObject(ServiceEndpoints, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
@@ -95,6 +119,12 @@ namespace Microsoft.Azure.Commands.Network.Models
         public string ServiceEndpointPoliciesText
         {
             get { return JsonConvert.SerializeObject(ServiceEndpointPolicies, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
+        }
+
+        [JsonIgnore]
+        public string InterfaceEndpointsText
+        {
+            get { return JsonConvert.SerializeObject(InterfaceEndpoints, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
         }
     }
 }